Fine-tune the update shell-script
Instead of using "git rev-parse", use "git show-ref" (which is the
proper way of getting the commit-hash for an active branch). When
reading the "revision"-file, whitespace are now discarded; the file is
now also (correctly) added to the commit to keep track of the current
merge-state. Finally, format a proper commit message.