Use Git post-checkout hook
Using Composer's post-install hook was causing all kinds of cross-
platform issues trying to run the shell-scrip. Git's post-checkout
hook happily executes "sh" on both Windows and Linux...

Required a check in the shell-script to prevent us running the automatic
update on master- and develop-branches. This in turn required a flag to
be added to force the script to run on the master-branch on its initial
run.

Updated README.md to reflect these changes.

PLOP-6